mlops/batch_inference/requirements.txt
apache-airflow==2.5.3
minio==7.1.14
pydantic==1.10.5
pandas==1.5.3
scikit_learn==1.2.2
